Use url variable in post error message instead of a global variable
A url (along with data) is provided as an parameter to the post
method. This url variable is then used for to post the data to the
webhook. If that fails, the resulting error message should name the
url it was trying to post to instead of the global POST_URL
variable (even if that variable is what gets passed to the post
method).