Add configuration for multiple webhook URLs
Add the configuration setting hooks.webhookurls to allow POSTing
to multiple webhooks instead of just a single one. The previous
hooks.webhookurl can still be used and the two can even be combined
where the hooks.webhookurl is the first webhook and the ones in
hooks.webhookurls are secondary.

The reason the splitting isn't added to the hooks.webhookurl is for
backwards compatibility where hooks.webhookurl might have a URL
that includes a comma which would then be split incorrectly.