Use 'empty tree' hash on get_revisions() if old param is 0000...
If <old> is 000... that means that the first commit ever on this git
repo is being pushed. In this case, using range such as
'%s~1..%s' % (new, new)
might result in error such as:
fatal: ambiguous argument 'e802996~1..e802996': unknown revision or path not in the working tree.
if new is the first commit of the repo.

By using the empty tree hash '4b825dc64...' it returns valid value.